API: restructure to match bluesky schema
