+++ tests/test_preprocessors.py
@@ -1,14 +1,11 @@
-import time
 import logging
 
 import pytest
-from bluesky.plans import scan
 from bluesky.plan_stubs import (trigger_and_read,
                                 create, read, save, null)
-from bluesky.preprocessors import run_decorator, run_wrapper
+from bluesky.preprocessors import run_decorator
 
-from pcdsdaq.plans import (daq_wrapper, daq_decorator, calib_cycle,
-                           calib_at_step)
+from pcdsdaq.preprocessors import daq_wrapper, daq_decorator
 
 logger = logging.getLogger(__name__)
